refactor: move the vtbackend input translation into input/
Key and mouse events on one side, the bytes written to the PTY on the other:
InputGenerator and its keyboard protocols, the InputHandler interface, the
InputBinding template a key binding is made of, and MatchModes -- the mode
predicate that decides whether a binding applies at all.

MatchModes belongs here rather than with the grid it appears to be about. Its
enumerators are AlternateScreen, AppCursor, AppKeypad, Insert, Select, Search and
Trace; its only structural consumer is InputBinding, and its only other callers
are the two key-binding sites in contour. It answers "does this chord fire", not
"what does the buffer look like".
