+++ examples/watch-mouse-events.cpp
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
-#include <vtbackend/InputGenerator.hpp>
 #include <vtbackend/core/Primitives.hpp>
+#include <vtbackend/input/InputGenerator.hpp>
 #include <vtbackend/vt/Sequence.hpp>
 
 #include <vtparser/Parser.hpp>
