+++ src/contour/input/MouseMapping.hpp
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 #pragma once
 
-#include <vtbackend/InputGenerator.hpp>
 #include <vtbackend/Terminal.hpp> // vtbackend::ScrollPhase
+#include <vtbackend/input/InputGenerator.hpp>
 
 #include <QtCore/Qt>
