+++ src/contour/session/TerminalSession.cpp
@@ -15,10 +15,10 @@
 #include <contour/session/TerminalSession.hpp>
 
 #include <vtbackend/HintModeHandler.hpp>
-#include <vtbackend/MatchModes.hpp>
 #include <vtbackend/Terminal.hpp>
 #include <vtbackend/ViCommands.hpp>
 #include <vtbackend/core/FileUrl.hpp>
+#include <vtbackend/input/MatchModes.hpp>
 
 #include <vtpty/Process.hpp>
 #include <vtpty/Pty.hpp>
