+++ src/vtbackend/Terminal.hpp
@@ -4,8 +4,6 @@
 #include <vtbackend/Cursor.hpp>
 #include <vtbackend/Folding.hpp>
 #include <vtbackend/HintModeHandler.hpp>
-#include <vtbackend/InputGenerator.hpp>
-#include <vtbackend/InputHandler.hpp>
 #include <vtbackend/Logging.hpp>
 #include <vtbackend/RenderBuffer.hpp>
 #include <vtbackend/Selector.hpp>
@@ -20,6 +18,8 @@
 #include <vtbackend/core/Hyperlink.hpp>
 #include <vtbackend/core/Primitives.hpp>
 #include <vtbackend/grid/Grid.hpp>
+#include <vtbackend/input/InputGenerator.hpp>
+#include <vtbackend/input/InputHandler.hpp>
 #include <vtbackend/vt/DesktopNotification.hpp>
 #include <vtbackend/vt/PointerShape.hpp>
 #include <vtbackend/vt/ProgressState.hpp>
